Trait fuzzy_matcher::FuzzyMatcher

pub trait FuzzyMatcher: Send + Sync {
    // Required method
    fn fuzzy_indices(
        &self,
        choice: &str,
        pattern: &str
    ) -> Option<(i64, Vec<usize>)>;

    // Provided method
    fn fuzzy_match(&self, choice: &str, pattern: &str) -> Option<i64> { ... }
}

Required Methods

fn fuzzy_indices( &self, choice: &str, pattern: &str ) -> Option<(i64, Vec<usize>)>

fuzzy match choice with pattern, and return the score & matched indices of characters

Provided Methods

fn fuzzy_match(&self, choice: &str, pattern: &str) -> Option<i64>

fuzzy match choice with pattern, and return the score of matching